cpplint fix

+++ src/Gui/Gui.cpp
@@ -42,8 +42,7 @@ void GLFWSetWindowCallback(GLFWwindow* window, GLFWkeyfun callback)
 
 void ImGuiSetStyle(ImGuiStyle& style)
 {
-  constexpr auto ColorFromBytes = [](uint8_t r, uint8_t g, uint8_t b)
-  { return ImVec4((float)r / 255.0f, (float)g / 255.0f, (float)b / 255.0f, 1.0f); };
+  constexpr auto ColorFromBytes = [](f32 r, f32 g, f32 b) { return ImVec4(r / 255.0f, g / 255.0f, b / 255.0f, 1.0f); };
 
   ImVec4* colors = style.Colors;
